Dealing with the split
SPlit playing should be proper, but not quite tested, as the final
end of play work needs to be finalized.

There's a new function for "end of hand" if split was invoked.  Need
to finish building that then test if it works.

If it does, then test if it scales.

Then build out the logic to automate.